remove unnecessary code duplication
